refactor: revert BackupInputFile split
