--- ksqldb-rest-app/src/main/java/io/confluent/ksql/rest/server/BackupInputFile.java
@@ -1,64 +1,0 @@
-/*
- * Copyright 2020 Confluent Inc.
- *
- * Licensed under the Confluent Community License (the "License"); you may not use
- * this file except in compliance with the License.  You may obtain a copy of the
- * License at
- *
- * http://www.confluent.io/confluent-community-license
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package io.confluent.ksql.rest.server;
-
-import io.confluent.ksql.util.Pair;
-import java.io.File;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-
-public class BackupInputFile {
-  protected static final String KEY_VALUE_SEPARATOR_STR = ":";
-  protected static final String NEW_LINE_SEPARATOR_STR = "\n";
-
-  static final byte[] KEY_VALUE_SEPARATOR_BYTES =
-      KEY_VALUE_SEPARATOR_STR.getBytes(StandardCharsets.UTF_8);
-  static final byte[] NEW_LINE_SEPARATOR_BYTES =
-      NEW_LINE_SEPARATOR_STR.getBytes(StandardCharsets.UTF_8);
-
-  private final File file;
-
-  public BackupInputFile(final File file) {
-    this.file = Objects.requireNonNull(file, "file");
-  }
-
-  public File getFile() {
-    return file;
-  }
-
-  public String getPath() {
-    return file.getAbsolutePath();
-  }
-
-  public List<Pair<byte[], byte[]>> readRecords() throws IOException {
-    final List<Pair<byte[], byte[]>> commands = new ArrayList<>();
-    for (final String line : Files.readAllLines(getFile().toPath(), StandardCharsets.UTF_8)) {
-      final String commandId = line.substring(0, line.indexOf(KEY_VALUE_SEPARATOR_STR));
-      final String command = line.substring(line.indexOf(KEY_VALUE_SEPARATOR_STR) + 1);
-
-      commands.add(new Pair<>(
-          commandId.getBytes(StandardCharsets.UTF_8),
-          command.getBytes(StandardCharsets.UTF_8)
-      ));
-    }
-
-    return commands;
-  }
-}
